Update init and Prog4
Updated init.py to make a text file that will populate the database with some fake information. I also updated Prog 4 to get the basic information needed for the actions we need to perform such as adding, deleting and updating. The queries are still in progress. The next steps are populating the database and coming up with the proper SQL statements that use the basic information retrieved.